Allow for very long ``uname -a'' lines.
On this machine it is "SunOS schubert.nada.kth.se 5.5.1
Generic_103640-05 sun4m sparc SUNW,SPARCstation-5"). Of course, this
causes GNAT's style checks to barf on s-garcon.ads.  For this reason,
the uname output is now wrapped with ``fmt'' (which hopefully is present
on all important Unices).

Subversion-branch: /importfromcvs/trunk
Subversion-revision: 46340