First draft of Broca.CDR whose ultimate goal is the replacement of Broca.Marshalling.
At this stage, the package has not been tested at all and does not
support
all CORBA types, nor is it efficient (I want to have a working
implementation
before I switch to efficient data structures that take advantage of the
alignment capabilities of CORBA).

However, any endianess is supported, including encapsulation with
different
endianess. The user will not have to manipulate the `Pos' counter as it
is hidden in the structure.

[imported from CVS by cvs2p4 at 2000/03/21 18:31:57]

[Imported from Perforce change 785 at 2006-12-01 18:28:57]

Subversion-branch: /trunk/adabroker
Subversion-revision: 31541