Started implementation of proper reference counting for all ojects. Please avoid using the Get primitive on CORBA.AbstractBase.Ref. Always use Object_Of instead.
No calls to Inc_Usage or Dec_Usage should be added to the code
from this point; the goal is to remove all existing calls
and to replace all pointers to objects derived from CORBA.Impl.Object
with references derived from CORBA.AbstractBase.Ref.

No new derivations of Broca.Refs.Ref or Broca.Refs.Entity should
be created, unless compelling reasons to do so arise. If new objects
need to be created, they should normally be derived from
CORBA.Impl.Object or one of its descendants.

This is not complete yet, but as far as it was tested this change
introduces no regressions. It is not meant for integration in
release 1.0.

[Imported from Perforce change 1689 at 2006-12-01 18:56:15]

Subversion-branch: /trunk/adabroker
Subversion-revision: 32231