Started specification of an abstract synch/asynch event handling facility.
---> Please review and comment <---

The intent is that Asynchronous_Event_Source will be implemented
as a set of file descriptors (within a derived type
Fd_Asynchronous_Event_Source). Wakeup will be implemented using
a UNIX pipe.

One Event_Manager will then be instanciated for the whole ORB,
and will be called by the Concurrency Management machinery to
provide work for ORB tasks.

One possible change to this specification would be to allow
a task to wait for events with certain specific properties.
Alternatively this can be implemented using multiple event
managers.

[Imported from Perforce change 2189 at 2006-12-01 19:10:31]

Subversion-branch: /trunk/adabroker
Subversion-revision: 32591