Simplify generated code by having it handle only Requests, not Messages. This allows the removal of one unit from the closure of s-polint.
Make RACW's From_Any a Renaming_TSS.

Introduce actual support for TypeCodes as TSS. This requires
a ugly hack to allow overloaded TSSs (a dummy access parameter is
added, just for the sake of name resolution, and a null value
is used in calls).

In the DSA example, recompile s-polint when necessary (assume
that the one installed in gnatlib is a dummy one).

[Imported from Perforce change 4709 at 2006-12-01 19:40:32]

Subversion-branch: /trunk/polyorb
Subversion-revision: 33601