Documentation cleanup.
2002/10/18 additional information:
A fix to PolyORB.Protocols.GIOP was also committed along the way.
It fixes a logic flaw whereby the wrong profile would be associated
to a Pending_Request.

The profile associated to a Pending_Request must be a profile
of the target reference of that request (and even /the/ profile
of that reference whose tag corresponds to the GIOP binding object
that is being used). The code up to change 5540 would use the
profile that was used to create the binding object. In the future,
when we reuse the same binding objects to communicate with different
objects hosted by the same remote middleware, this would have
become incorrect.

[Imported from Perforce change 5540 at 2006-12-01 19:48:02]

Subversion-branch: /trunk/polyorb
Subversion-revision: 33944