New implementation of server side of remote access-to-subprogram types. We now create a proxy object for each RCI subprogram on which we can make dispatching calls through the RACW type associated with any appropriate RAS.
The next steps are removal of the stub for the case of local
calls (having the RAS designate the proxy object directly,
when there is no All_Calls_Remote pragma), and then removal
of the legacy fat type (replacing it with the RACW type).

Recompute build deps.

Part of work towards integration with GLADE under CA22-003.

[Imported from Perforce change 8005 at 2006-12-01 20:28:46]

Subversion-branch: /trunk/polyorb
Subversion-revision: 35621