Refactor Chained_Lists to allow for optional double chaining, controlled by a generic formal parameter.
It is now mandatory to use double chaining if removal or
insertion of list elements at an arbitrary iterator position
is desired. Removal of elements that verify a certain predicate
does not require double chaining, and is now provided through
a generic procedure Remove_G.

For D708-012.

[Imported from Perforce change 8347 at 2006-12-01 20:33:04]

Subversion-branch: /trunk/polyorb
Subversion-revision: 35945