When a scoped name denotes an interface, it is necessary to check whether the interface declaration occurs within the current gen_scope, or any subscope thereof. If it is the case, the reference must be mapped to a reference to a forward interface, possibly generated implicitly for that purpose.
The previous implementation of this check did not correctly account
for the fact that the current scope might be several levels deeper
than the current gen_scope, depending on the presence of nested
structures. It also did not take the case of unions into account
altogether. This new implementation addresses both shortcomings.

Fixes E419-009.

[Imported from Perforce change 9277 at 2006-12-01 21:17:57]

Subversion-branch: /trunk/polyorb
Subversion-revision: 36784