Add new configuration mechanisms ie command line configuration and partition unit configuration. The latter is supposed to be a configuration unit automatically generated by a configuration tool such as gnatdist. To evaluate a configuration variable, all the configuration mechanisms are considered and the first configuration mechanism matching the n-uple (section, key) is selected. Using the basic configuration setup, the configurations mechanisms are checked in the following order : command line, environment variable, configuration file, partition unit.
PolyORB allows to set up configuration variables on the
command line. The syntax is close to the one described in
PolyORB.Parameters.File or PolyORB.Parameters.Environment. A
variable Var.Iable in section [Sec] can be overriden with flag
--polyorb-<sec>-<var>-<iable>[=<value>]. If no value is
provided, then the returned value is set to "true" since we
consider that the feature is activated. Note that Section and
Key have to be in lower case.

Define a dummy partition configuration unit which is imported
in PolyORB.Setup.Base. The objective is to allow to redefine
this unit locally to a given partition with a configuration
tool and automatically import this unit by withing
PolyORB.Setup.Base.

Simplify some tasking configuration variable names since they
were too long and duplicating somehow the Section part. Fix the
units using them as well as the configuration file.

Define a base configuration setup.

[Imported from Perforce change 9504 at 2006-12-01 21:45:38]

Subversion-branch: /trunk/polyorb
Subversion-revision: 36987