In the case of local requests, both client stubs and server skels declare objects for parameters, and provide foreign Any's wrapping these objects. In Arguments on server side when reconciling the initial argument list of the request with the one provided by the servant, we may not assume that either list contains empty Any's; in the case of both lists providing non-empty Any's, we need to assign the underlying values.
Similarly for results, we may not assume that the Any seen by
the server skel is empty (as in the case of a local call it is
the one provided by the client stubs, which wraps the temporary
result object) and we need to perform an assignment.

This change implements in-place value assignment in
PolyORB.Any.Copy_Any_Value for the case where Dst is not empty; to
this end, the Clone operation on PolyORB.Any.Content is extended
to accept a second parameter which, if non-null, is a content on
which in-place assignment must be performed.

Followup on EB18-024

Fixes regression on test TIME_0 from the test suite.

[Imported from Perforce change 10243 at 2006-12-01 22:52:17]

Subversion-branch: /trunk/polyorb
Subversion-revision: 37704