New implementation of sequences (both bounded and unbounded) with reduced code footprint. The code is now split in two parts: * PolyORB.Sequences contains all the complicated computation of   slice indices to be moved around for each operation; these   moves are represented as 'programs' that define a sequence of   slice assignments;
* Each instance of PolyORB.Sequences.(Bounded|Unbounded) just
  provides a virtual machine that can execute these programs on
  a particular type of sequence.

This reimplementation was prompted by the necessity of reducing
the code footprint for each instance of sequences (FA17-009).

Since the spec of PolyORB.Sequences.* is derived from the bounded
and unbounded string packages of the standard Ada runtime library,
this implementation was tested against the relevant ACATS tests.

See FB24-019.

[Imported from Perforce change 10469 at 2006-12-01 22:56:13]

Subversion-branch: /trunk/polyorb
Subversion-revision: 37928