Reorganization of the exception handling code for CORBA:
Move management of the AdaExceptionInformation service context
from CORBA layer to neutral layer (so that it can be shared with
other personalities).

(PolyORB.CORBA_P.Exceptions.Request_Raise_Exception): New subprogram,
similar in function (but not identical) to the identically named
procedure in the DSA personality. Raises the current exception
occurrence for the given request, if any. This allows code to be
factored
across all client stubs.

Cleanup done as part of G228-024

Subversion-branch: /trunk/polyorb
Subversion-revision: 42897