Follow-on to r43058.
Make it work in case the build area is separate from the source area.

In particular: ADA_PROJECT_PATH needs to include the projects
subdirectory of
both areas, so we can find src_setup.gpr. The Source_Dirs of
src_setup.gpr.in
needs to look in both areas, since source dirs are relative to the
directory
containing the project file.

Note: "make src_setup.gpr" will no longer work. I put in a ??? comment
for that
in Makefile.in.

Subversion-branch: /trunk/polyorb
Subversion-revision: 43063