Improve efficiency of typecodes handling.
TypeCodes have been made into reference-counted entities for proper
garbarge collection as part of G405-028. However, within representation
functions implementations (e.g. PolyORB.Representations.CDR), we know
that
we globally do not change the reference count of typecode objects when
we
are using them to marshall and unmarshall data, and so it is inefficient
to create and destroy full-fledged smart pointers to typecode objects in
such contexts. Instead, we now provide direct accessors for the internal
structure of typecodes that expose normal pointers, and do not do any
reference counting.

We also disable reference counting altogether for typecodes statically
generated at libray level, since we know that we do not need or want
these
to be garbage collected.

Subversion-branch: /trunk/polyorb
Subversion-revision: 43865