Reorganize project files generation so that visibility on the PCS is established using project files instead of (in the case of PolyORB) relying on an external shell script (polyorb-config), which can't be used on Windows.
For now, using polyorb-config is still the default on UNIX platforms,
but the Windows behaviour (use project file instead) can be requested
using debugging switch -dP.

To this effect, the main distributed application project file
(dist_app_project) is made an "extends all" of the user project,
which includes a dependency on the PCS. This ensures that when
user sources are recompiled as part of the dist_app_project, they
have visibility on the PCS. This also requires that the dist_app_project
provides an object directory of its own, dsa/obj. All operations that
need to access the monolithic app's ALI files are adjusted accordingly.

The list of PCS units that are overridden by a given partition depends
on the partition (in the GARLIC case, s-gaprco and s-gastco may or may
not be overridden, and when not, the default PCS version must be used).
This is supported by generating a list of excluded PCS units for each
partition, which is referenced through an external variable in the PCS
project.

This fixes H313-028 (PolyORB/DSA support on Windows).

Subversion-branch: /trunk/polyorb
Subversion-revision: 126420