Reorganization of autoconf setup.
This change removes the need for the prepare_distrib script to modify
configure.ac. Instead, when preparing a source package, an additional
file is created (support/distrib.m4) which contains appropriate
definitions
to tweak those settings that need to be changed from their default value
(which is used when building straight from an SCM checkout).

Subversion-branch: /trunk/polyorb
Subversion-revision: 127140