Implement support for RPC abortion in GIOP and neutral layer. An RPC can now be aborted on the client side, and this causes a notification to be sent on the session, and an attempt to abort execution on the server side.
When a client disconnects, an abort is also implicitly generated for all
its
requests that are still being served.

Fixes J817-001


----
date: 2010/09/03 17:36:41;  author: quinot;
Change committed by mistake as part of a larger commit.
RH should have read:
Fix to change for J521-019 to also pass name_service on the command line
in the case of a local launch, since we can't rely on the parameter
always
being present in the environment.
----
date: 2010/09/03 17:37:23;  author: quinot;
(the above comment applies to the modification of
polyorb-dsa_p-remote_launch.adb in this revision).

Subversion-branch: /trunk/polyorb
Subversion-revision: 164266