(PolyORB.Asynch_Ev.Stabilize): New operation of event handlers, returns True if we can guarantee that the object on which the event is occurring won't disappear while we are processing the event.
(PolyORB.Transport.Handlers.Stabilize): For a transport endpoint,
stabilization is achieved by setting a reference to the corresponding
binding object, which will be kept until the event processing is
completed.
(PolyORB.Transport.Handlers.Handle_Event): Keep a reference to the
dependent
BO set in the handler.

(PolyORB.ORB.Try_Check_Sources): Do not attempt to process events on an
event handler that cannot be stabilized, i.e. that corresponds to a
binding
object that is being torn down.

Fixes M313-014

Subversion-branch: /trunk/polyorb
Subversion-revision: 204265