copy testsuite.py as legacy_py2_testsuite.py
The purpose of this allow us to keep the Python-2.x version of
this testsuite while we work on the next version of the testsuite
framework, to be implemented using Python 3.x.

no-precommit-check: The script has style violations that are not worth
fixing, since we'll eventually abandon it.

TN: U922-064
Change-Id: I8cdee8309ecfec770ddaa99e74dc5f720b609795