switch the documentation's them to the RTD them
This commit adjust the sphinx configuration to use the "Read The Docs"
theme, which has the advantage of allowing the navigation bar
(containing among other things a search bar, and the TOC) to stay
fixed while scrolling the contents of the page being read. This is
particularly useful to allow access to those features while reading
a long page, for instance.

TN: VB25-026
Change-Id: Ibfebbf1d66ed34e72c4f17a90f0272ed7af787b7